+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Detect preferred color scheme.
 - Show loading indicator while attachment is being uploaded.
+- Content warnings.
 
 ### Fixed